Expand caching decorators to accommodate invalidation
While caching resources identifiable by only a single identifier is
useful, it's not as powerful as it could be. Namely, we can cache search
operations by a given set of arguments, and invalidate likewise, which
means we can now do more than just cache, say, retrieve operations.

Additionally, we can now have a dedicated way of invalidating said
resources based on known query args. At the time of validation, we'll
adopt a convention where keys are "locked" on deletion, and incoming
requests won't be able to set until the deleted key reaches its TTL.

There will thus need to be some care taken with which
endpoints/resources are cached. If, for example, the resource is being
searched via any given number of arbitrary arguments, then you would
need to account for all permutations. This could probably benefit from a
whitelist of arguments.

Note: This works via microcosm conventions, that incoming controllers are
always passed via kwargs. This could probably be expanded via inspection,
but I'm choosing to kick that can down the road for the moment given
that it won't actually be used by any service we run today.