Replace hash() with sha1() for key hashing
While `hash()` is a convenient hashing mechanism, it has the unfortunate
side-effect of making the hashing key non-deterministic across
processes. This means that different instances caching the same
resource/arguments may end up using different keys.

This change switches to using `sha1()`. This is similar to how we've
cached resources in the past (e.g. via `uuid5`), and should be
deterministic.