GLOB-39083 Update cache key logic (#12)
One issue with the previous implementation was that if a service
changes (either altering the resource definition, cached function, etc.)
we'd still try and pull the cached resource and attempt to treat it as
if it's valid, when there's no guarantee that that's the case.

We should be able to thus use the commit SHA as a proxy for a cache
"version". This will have the side-effect of causing each new build to
invalidate its entire cache, but will be safer by default.

Note that this is probably the wrong abstraction: this idea was pulled
from microcosm-flask, but could probably be better served elsewhere (for
both usages).